fix minor UTF8 bug in yaml-to-latex.py
If the user enters non-ascii metadata in metadata.yaml in UTF8,
e.g.
    name:    Institute of Ąćçęńted Letterś in Półiśh and Freńćh
    address: Żyźszczyń
but in the terminal has a non-UTF8 locale setting such as LANG=C
or LANG=fr_FR.ISO-8859-1, then yaml-to-latex.py will yield an
error and 'make' will fail.

This commit sets the locale, within the python process, to en-US.UTF-8.

Further improvements could be to check whether the existing locale
is UTF-8 and to warn the user if it is not.