DOI: oadoi.org; ArXiv identifiers; URL display
This commit changes the DOI links in the references to open
access DOI oadoi.org for the link and displays the DOI identifier
in the text. Related changes: \sloppy option for references is
inserted, to avoid having text extending into the margin;
(most) URL texts are made copy/pastable by redefining \nolinkurl
to print text without alteration.

The text appearance of the eprint identifers was wrong because
it inserted a space into the identifier; see
https://arxiv.org/help/arxiv_identifier . This is fixed in
this commit.

The display of URLs is enabled in this commit. Linking to the URL
archives is good, but the original URLs should be shown too -
archives can themselves be sometimes unavailable, not only URLs.

TODO: A better version of the URL fix still needs to be considered.
(Examples where the text URL is unusable in Roukema 2020
ReScience C, 6, 1, #11 = https://zenodo.org/record/3956058 are
ref [12] with
`https : / / techcrunch . com / 2019 / 07 / 29 / github - ban -`
and ref [24] with `https : / / gcc . gnu . org /`.)

* modified:   rescience.cls

This commit was cherry-picked from the original d4cc1fc.